+++ module/Core/src/Visit/Repository/VisitDeleterRepository.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Core\Visit\Repository;
+
+use Happyr\DoctrineSpecification\Repository\EntitySpecificationRepository;
+use Shlinkio\Shlink\Core\ShortUrl\Entity\ShortUrl;
+use Shlinkio\Shlink\Core\Visit\Entity\Visit;
+
+class VisitDeleterRepository extends EntitySpecificationRepository implements VisitDeleterRepositoryInterface
+{
+    public function deleteShortUrlVisits(ShortUrl $shortUrl): int
+    {
+        $qb = $this->getEntityManager()->createQueryBuilder();
+        $qb->delete(Visit::class, 'v')
+           ->where($qb->expr()->eq('v.shortUrl', ':shortUrl'))
+           ->setParameter('shortUrl', $shortUrl);
+
+        return $qb->getQuery()->execute();
+    }
+}