+++ module/Core/src/Visit/Repository/VisitDeleterRepositoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Core\Visit\Repository;
+
+use Shlinkio\Shlink\Core\ShortUrl\Entity\ShortUrl;
+
+interface VisitDeleterRepositoryInterface
+{
+    public function deleteShortUrlVisits(ShortUrl $shortUrl): int;
+}